Remove weak_rng
Documentation for the function claims it will return the fastest algorithm
available in Rust, but since its return type is hardcoded to XorShiftRng, it
would be an impossible [breaking-change], when, or if, a faster Rng is added to
Rust.

[breaking-change], since this removes public API. Uses of `weak_rng` function
can be replaced with `random::<XorShiftRng>()` to achieve almost the same
behaviour.