Replace the `IndependentSample` trait with `Distribution`.
Also removed the `Sample` trait.

`IndependentSample` is really a probability distribution of values, and hence renamed to `Distribution`, but with an associated `Output` type instead of being generic over a type.  The `Sample` trait is a state-mutating version of `IndependentSample` and is really a random process.  A random process can be infinite (e.g. a random walk) or finite (e.g. drawing from a bag without replacement).  So really, a random process is best represented by an `Iterator`.  Additionally all previous implementations of `Sample` simply called through to `IndependentSample.ind_sample`, so `Sample` has simply been removed.

The related trait `SampleRange` has been renamed to `RangeDistribution`.