Fix iterator protocol issue in rand::sample
It's unspecified what an iterator does after it has returned None the
first time, so we have to check how many elements we have afte the
initial extraction in `sample`.

Also modernize sample by using `IntoIterator`.

The type parameter `T` is redundant, but removing it would be a breaking
change.