[breaking change] Automatically impl Sample for T: IndependentSample
This removes the need to manually implement Sample for types that
implement IndependentSample, cutting down on boilerplate and reducing
chances for mis-matched implementations.

It also adds blanket impls for `&T` where `T: IndependentSample`, so
that `.sample()` can be called on these types without needing a unique
reference.

This is a breaking change because it conflicts with crates that impl
both Sample and IndependentSample for their own types.