Change derive behavior for enums with < 3 variants
Enums with a single variant will simply have the inner type (if any)
generated. This avoids extra calls on the Rng.

Enums with two variants will generate a bool rather than a number in
a range. This should be a much faster approach, although I have not
tested it.

These changes avoid calls to the unreachable!() macro, which generates
a panic in unnecessary cases like these. This in turn may make the
generated functions faster since unused instructions won't be loaded.

Enums with 3 or more variants behave the same as before.