Add `new_from_u64` to `IsaacRng` and `Isaac64Rng` I have implemented it as a function instead of a trait, as that makes it easy to add it to every RNG ont at a time.
I split the `init` function in two instead of the current version that uses a
bool to select between two paths. This makes it more clear how the seed is used.
The current `mix` macro has to be defined in the function, and would have to be
duplicated. Therefore I converted it to a seperate function.

I precalculated the values a...h, but am not sure this is a good idea. It makes
the resulting code smaller, and gives a small performance win. Because it are
'magic' values anyway, I thought why not?