Switch rand_os to use fuchsia-cprng crate
The fuchsia team has spun out a new crate,
[fuchsia-cprng](https://crates.io/crates/fuchsia-cprng), which reduces
your exposure to changes in our syscalls, which are still not stable.
This crate just exposes the CPRNG syscall, which we do not expect to
change.

Also note, the underlying syscall,
[zx_cprng_draw](https://fuchsia.googlesource.com/zircon/+/HEAD/docs/syscalls/cprng_draw.md)
no longer needs the loop. Instead, we guarantee that the buffer always
be filled with randomness.