Remove unnecessary import of fuchsia-zircon
The fuchsia project is trying to purge users of fuchsia-zircon
and instead replace them with the much slimmer fuchsia-cprng,
which only exposes the CPRNG for the OS. There are still a few
users of 0.3.x, like multipart and rust-crypto. Since rand 0.3
is just re-exporting 0.4, there is no reason to explicitly
depend on fuchsia-zircon, so this just removes this dependency.