rand: Remove serde1 feature
This feature does not do anything in the `rand` crate and was forwarded
to `rand_core`, where it was used for `BlockRng` and `BlockRng64`.
However, crates using this feature would use `rand_core` directly
instead, making the feature pointless for `rand`.

It is not expected to use this feature for any of the RNGs in `rand`.

This hopefully makes it possible to build `rand` on docs.rs again.
(See https://github.com/rust-lang/docs.rs/issues/369.)

Now the serde1 feature does not do anything, but it is still available
for backwards compatibility.