Try to clarify the documentation of LogNormal::new
It may be that I am the only one who expected the 'mean' parameter of LogNormal::new() to be the actual mean of the resulting LogNormal distribution. :-) Reading up a bit more on the subject, I realize that the definition used is probably reasonable, but I thought maybe we could clarify it a bit?

Feel free to just disregard this PR if you don't think the clarification is necessary (or if it's wrong).