Use `f32` arithmetic for `f32` normal distribution
(As suggested in the TODO.)

This is faster on 32-bit platforms instead of computing  the `f64` and
downconverting.

In addition to the tests, I also generated some normal numbers with the
current f64 and the new f32 paths and ensured that they both look
"normal".

I'm very tempted to convert the ziggurat tables to `f32` for all versions:
they are only used when multiplying against against a random,
full-precision float. So, for `f64`, we'd still be multiplying a full
52-bit mantissa against the 23-bit mantissa of the table, which should
be sufficient for generating a random `f64` with full precision. And
then we'd save 50% of the space of the table. Thoughts?